Не работает
Добавил метод contains. Он почему-то не может проверить список фигур.

Попробова вынести фигуры paintComponent. Теперь они не добавляются к
списку

+++ src/Point.java
@@ -5,9 +5,9 @@ public class Point extends Vector {
     int x;
     int y;
 
-    public Point(int x, int y) {
-        this.x =  x;
-        this.y =  y;
+    public Point(double x, double y) {
+        this.x =  (int) x;
+        this.y =  (int) y;
     }
 
     public Point() {
